Adding all tutorials pages with illustration
