+++ app/src/main/java/com/sleewell/sleewell/mvp/help/OnBoardingActivity.kt
@@ -1,9 +1,12 @@
 package com.sleewell.sleewell.mvp.help
 
+import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.cuberto.liquid_swipe.LiquidPager
 import com.sleewell.sleewell.R
+import com.sleewell.sleewell.modules.settings.SettingsManager
+import com.sleewell.sleewell.mvp.mainActivity.view.MainActivity
 
 class OnBoardingActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -12,5 +15,13 @@ class OnBoardingActivity : AppCompatActivity() {
 
         val pager = findViewById<LiquidPager>(R.id.pager)
         pager.adapter = OnBoardingAdapter(supportFragmentManager)
+    }
+
+    fun dismissActivity() {
+        val settings = SettingsManager(this)
+
+        settings.setTutorial(false)
+        val intent = Intent(this, MainActivity::class.java)
+        startActivity(intent)
     }
 }
