+++ app/src/main/java/com/sleewell/sleewell/mvp/help/onBoardingPages/PageAlarmFragment.kt
@@ -1,0 +1,20 @@
+package com.sleewell.sleewell.mvp.help.onBoardingPages
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.sleewell.sleewell.R
+
+
+class PageAlarmFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_page_alarm, container, false)
+    }
+}
