+++ app/src/main/res/drawable/undraw_mic_drop_uuyg.xml
@@ -1,0 +1,54 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="488dp"
+    android:height="692.3065dp"
+    android:viewportWidth="488"
+    android:viewportHeight="692.3065">
+  <path
+      android:pathData="M0,627.759h488v2h-488z"
+      android:fillColor="@color/colorPrimaryDark"/>
+  <path
+      android:pathData="M313.616,192.464s-20.74,3.453 -31.397,1.516 -31.801,9.056 -31.801,9.056L194.541,223.724S144,265.134 136.893,236.281C132.912,220.122 185,207.134 185,207.134l91.555,-43.856L309.868,151.764Z"
+      android:fillColor="#ffb9b9"/>
+  <path
+      android:pathData="M317.606,313.134l-14.606,14 2.606,61s3,80 2,83 -19,125 -19,125 21,10 40,4c0,0 19,-90 19,-116l22,-91 34,104s2,70 6,93 4,26 4,26 32,9 40,0c0,0 -4,-118 -7,-130l-7,-143Z"
+      android:fillColor="@color/colorPrimaryDark"/>
+  <path
+      android:pathData="M296.606,596.134s-2,-3 -8,5 -25,23 -25,23 -52,37 -13,39 46,-18 46,-18 33,-6 34,-13 -4,-35 -4,-35Z"
+      android:fillColor="@color/colorPrimaryDark"/>
+  <path
+      android:pathData="M415.606,614.134s-7,24 -5,34 -4,47 24,44 19,-35 19,-35l-9,-41Z"
+      android:fillColor="@color/colorPrimaryDark"/>
+  <path
+      android:pathData="M376.106,57.634m-33,0a33,33 0,1 1,66 0a33,33 0,1 1,-66 0"
+      android:fillColor="#ffb9b9"/>
+  <path
+      android:pathData="M356.606,75.134s10,35 2,48 54,2 58,-10c0,0 -20,-22 -17,-43Z"
+      android:fillColor="#ffb9b9"/>
+  <path
+      android:pathData="M372.606,122.134s27.374,1.851 40.687,-13.075l35.616,15.221a18.685,18.685 0,0 1,11.043 20.51L441.606,246.134s8.394,64 4,107c-3.97,38.849 -148,0 -143,-33 0.536,-3.536 1.03,-7.152 1.486,-10.79a851.217,851.217 0,0 1,18.207 -96.54c2.659,-10.864 -4.693,-72.67 -4.693,-72.67l44.018,-29.165S363.606,122.134 372.606,122.134Z"
+      android:fillColor="@color/colorAccent"/>
+  <path
+      android:pathData="M342.931,137.548l-8.368,-8.601 -61.292,34.515s17.955,31.363 25.025,31.46c2.71,0.037 34.866,1.61 42.703,-8.789C353.611,169.4 342.931,137.548 342.931,137.548Z"
+      android:fillColor="@color/colorAccent"/>
+  <path
+      android:pathData="M367.999,36.041a131.948,131.948 0,0 1,14.787 -1.466c3.588,-0.152 7.327,-0.128 10.569,1.416 4.889,2.328 7.629,7.592 9.422,12.702s3.066,10.57 6.393,14.843c5.536,-5.778 6.642,-14.369 7.466,-22.328 0.58,-5.598 1.094,-11.568 -1.495,-16.566 -1.594,-3.078 -4.225,-5.477 -6.889,-7.695 -12.72,-10.587 -28.875,-18.917 -45.252,-16.539l-0.383,2.528a45.156,45.156 0,0 0,-8.354 0.525l0.585,6.273 -8.699,0.552 -0.334,6.538 -8.797,0.305a10.193,10.193 0,0 0,3.069 6.627,12.392 12.392,0 0,1 -4.519,2.041c1.467,5.679 2.891,11.371 4.407,17.037 1.116,4.174 0.917,7.862 5.394,5.825 3.409,-1.551 5.715,-6.483 9.161,-8.521C358.59,37.738 363.395,36.761 367.999,36.041Z"
+      android:fillColor="@color/colorPrimaryDark"/>
+  <path
+      android:pathData="M476.339,181.745l0.485,35.243L466.672,317.996s5.476,53.402 -11.074,55.147c-29.552,3.115 -7.77,-58.486 -7.77,-58.486l0.479,-59.582s3.158,-23.622 -2.284,-32.987 -9.238,-30.045 -9.238,-30.045Z"
+      android:fillColor="#ffb9b9"/>
+  <path
+      android:pathData="M448.606,127.134h0a18.678,18.678 0,0 1,17.95 13.495c6.386,22.129 17.985,63.771 15.05,64.505 -4,1 -51,4 -51,4Z"
+      android:fillColor="@color/colorAccent"/>
+  <path
+      android:pathData="M83.626,296.796l0.241,0.058a21.334,21.334 0,0 1,3.279 40.399,21.002 21.002,0 0,1 -5.088,1.436 21.341,21.341 0,1 1,1.569 -41.893Z"
+      android:fillColor="@color/colorAccent"/>
+  <path
+      android:pathData="M83.626,296.796a306.277,306.277 0,0 1,3.52 40.457,21.002 21.002,0 0,1 -5.088,1.436 222.047,222.047 0,0 0,-3.707 -42.432A21.308,21.308 0,0 1,83.626 296.796Z"
+      android:fillColor="#e6e6e6"/>
+  <path
+      android:pathData="M168.563,301.251l-73.204,2.971 2.062,24.232 72.752,-10.973a4.991,4.991 0,0 0,4.275 -5.514l-0.099,-0.856 2.872,-0.119a2.521,2.521 0,0 0,2.391 -2.88l-0.018,-0.122a2.521,2.521 0,0 0,-2.762 -2.146l-3.052,0.325 -0.058,-0.502A4.991,4.991 0,0 0,168.563 301.251Z"
+      android:fillColor="@color/colorPrimaryDark"/>
+  <path
+      android:pathData="M128.014,315.517a1.487,1.487 0,0 1,-1.385 1.138l-5.258,0.218a1.495,1.495 0,0 1,-1.526 -1.219l-0.533,-2.905a1.489,1.489 0,0 1,1.307 -1.749l5.32,-0.567a1.484,1.484 0,0 1,1.631 1.267l-0,-0 0.471,3.255A1.494,1.494 0,0 1,128.014 315.517ZM120.293,312.363a0.495,0.495 0,0 0,-0.005 0.206l0.533,2.906a0.498,0.498 0,0 0,0.509 0.406l5.258,-0.218a0.496,0.496 0,0 0,0.471 -0.567l-0.471,-3.255a0.494,0.494 0,0 0,-0.543 -0.422l-5.32,0.567a0.489,0.489 0,0 0,-0.345 0.196A0.494,0.494 0,0 0,120.293 312.363Z"
+      android:fillColor="#e6e6e6"/>
+</vector>
