+++ app/src/main/res/drawable/undraw_mornings_re_cofi.xml
@@ -1,0 +1,60 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="1082.096dp"
+    android:height="583.577dp"
+    android:viewportWidth="1082.096"
+    android:viewportHeight="583.577">
+  <path
+      android:pathData="M873.57,583.58L217.45,583.58c-6.71,-0.01 -12.14,-4.61 -12.15,-10.28v-474.77c0.01,-5.68 5.44,-10.27 12.15,-10.28L873.57,88.25c6.71,0.01 12.14,4.61 12.15,10.28v474.77C885.72,578.98 880.28,583.58 873.57,583.58ZM217.45,92.36c-4.02,0 -7.29,2.76 -7.29,6.17v474.77c0,3.4 3.27,6.16 7.29,6.17L873.57,579.47c4.02,0 7.29,-2.76 7.29,-6.17v-474.77c0,-3.4 -3.27,-6.16 -7.29,-6.17Z"
+      android:fillColor="#e6e6e6"/>
+  <path
+      android:pathData="M487.31,281.09l175,-34.99 3.8,96.01 -13.7,108.76 9.9,129.71L473.57,580.58l4.91,-150.49s-42.11,-59.42 -17.3,-119.31Z"
+      android:fillColor="@color/colorAccent"/>
+  <path
+      android:pathData="M877.52,53.33a28.21,28.21 0,0 1,-26.5 34.19l-167.81,194.6 -39.02,-47.35 177.22,-176.39a28.36,28.36 0,0 1,56.11 -5.05Z"
+      android:fillColor="#feb8b8"/>
+  <path
+      android:pathData="M719.74,260.88a12.64,12.64 0,0 1,-4.18 9.67l-47.19,42.48a35.13,35.13 0,1 1,-54.3 -44.6l32.94,-54.51a12.65,12.65 0,0 1,18.89 -3.21l49.25,40.7a12.64,12.64 0,0 1,4.59 9.48Z"
+      android:fillColor="@color/colorAccent"/>
+  <path
+      android:pathData="M212.91,510.64a28.21,28.21 0,0 1,31.33 -29.83l195.11,-167.21 31.47,52.67 -201.68,147.8a28.36,28.36 0,0 1,-56.23 -3.43Z"
+      android:fillColor="#feb8b8"/>
+  <path
+      android:pathData="M568.63,126.81m-71.24,0a71.24,71.24 0,1 1,142.48 0a71.24,71.24 0,1 1,-142.48 0"
+      android:fillColor="@color/colorSecondary"/>
+  <path
+      android:pathData="M400.04,329.11a12.64,12.64 0,0 1,5.58 -8.93l53.03,-34.92a35.13,35.13 0,1 1,47.29 51.97q-0.15,0.14 -0.3,0.27l-40.75,48.96a12.65,12.65 0,0 1,-19.15 0.34l-42.59,-47.63a12.64,12.64 0,0 1,-3.12 -10.06Z"
+      android:fillColor="@color/colorAccent"/>
+  <path
+      android:pathData="M561.89,171.79m-64.5,0a64.5,64.5 0,1 1,129 0a64.5,64.5 0,1 1,-129 0"
+      android:fillColor="#feb8b8"/>
+  <path
+      android:pathData="M605.82,0a38.94,38.94 0,0 1,-22.47 15.45c-15.71,4.21 -0.13,27.82 -0.13,27.82s-5.67,-8.59 2.51,-17.68S607.05,10.28 605.82,0Z"
+      android:fillColor="@color/colorSecondary"/>
+  <path
+      android:pathData="M580.43,48.93m-19.85,0a19.85,19.85 0,1 1,39.71 0a19.85,19.85 0,1 1,-39.71 0"
+      android:fillColor="@color/colorSecondary"/>
+  <path
+      android:pathData="M610.1,65.57c-16.19,-6.12 -43.14,-9.23 -57,-7.1l1.9,-13.89c13.33,6.89 40.16,10.13 57,7.1Z"
+      android:fillColor="@color/colorAccent"/>
+  <path
+      android:pathData="M572.7,68.17s-20.28,33.74 -67.45,46.37 -0.38,83.51 -0.38,83.51 -17.01,-25.77 7.53,-53.07S576.42,99.02 572.7,68.17Z"
+      android:fillColor="@color/colorSecondary"/>
+  <path
+      android:pathData="M580.39,68.49s19.49,46.34 40.88,50.73 19.77,77.61 -13.69,88.59c0,0 24.4,-42.93 3.54,-71.72S577.11,84.54 580.39,68.49Z"
+      android:fillColor="@color/colorSecondary"/>
+  <path
+      android:pathData="M0,580.58h1082.1v2h-1082.1z"
+      android:fillColor="#e6e6e6"/>
+  <path
+      android:pathData="M188.88,533.05a19.09,19.09 0,0 0,-19.09 -19.09h-0v-14.68L96.37,499.28v72.66a9.56,9.56 0,0 0,9.56 9.56h54.29a9.56,9.56 0,0 0,9.56 -9.56L169.79,552.13A19.09,19.09 0,0 0,188.88 533.05ZM169.79,544.79 L169.79,544.79v-23.49a11.75,11.75 0,0 1,0 23.49Z"
+      android:fillColor="#e6e6e6"/>
+  <path
+      android:pathData="M973.47,511.72c14.38,15.43 15.12,38.1 15.12,38.1s-22.57,-2.34 -36.94,-17.76 -15.12,-38.1 -15.12,-38.1S959.09,496.29 973.47,511.72Z"
+      android:fillColor="#e6e6e6"/>
+  <path
+      android:pathData="M1018.72,535.8c-15.43,14.38 -38.1,15.12 -38.1,15.12s2.34,-22.57 17.76,-36.94 38.1,-15.12 38.1,-15.12S1034.15,521.42 1018.72,535.8Z"
+      android:fillColor="#e6e6e6"/>
+  <path
+      android:pathData="M1014.89,581.58h-56.74a5.8,5.8 0,0 1,-5.22 -3.27l-12.35,-25.4a5.8,5.8 0,0 1,5.22 -8.34h81.45a5.8,5.8 0,0 1,5.22 8.34l-12.35,25.39a5.8,5.8 0,0 1,-5.22 3.27Z"
+      android:fillColor="#e5e5e5"/>
+</vector>
