+++ app/src/main/res/drawable/undraw_personal_settings_re_i6w4.xml
@@ -1,0 +1,108 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="676.6583dp"
+    android:height="577.3477dp"
+    android:viewportWidth="676.6583"
+    android:viewportHeight="577.3477">
+  <path
+      android:pathData="M315.043,528.398a24.215,24.215 0,0 0,23.383 -4.119c8.19,-6.874 10.758,-18.196 12.847,-28.682l6.18,-31.017 -12.938,8.908c-9.305,6.406 -18.818,13.019 -25.26,22.298s-9.252,21.947 -4.078,31.988"
+      android:fillColor="#e4e4e4"/>
+  <path
+      android:pathData="M317.041,568.108c-1.628,-11.864 -3.304,-23.881 -2.159,-35.872 1.015,-10.649 4.264,-21.049 10.878,-29.579a49.206,49.206 0,0 1,12.625 -11.44c1.262,-0.796 2.424,1.204 1.167,1.997A46.779,46.779 0,0 0,321.048 515.539c-4.029,10.246 -4.675,21.416 -3.982,32.3 0.419,6.582 1.311,13.121 2.206,19.653a1.198,1.198 0,0 1,-0.808 1.423,1.163 1.163,0 0,1 -1.423,-0.808Z"
+      android:fillColor="#f2f2f2"/>
+  <path
+      android:pathData="M328.762,549.072a17.825,17.825 0,0 0,15.531 8.019c7.864,-0.373 14.418,-5.86 20.317,-11.07l17.452,-15.409 -11.55,-0.553c-8.306,-0.398 -16.827,-0.771 -24.738,1.793s-15.208,8.726 -16.654,16.915"
+      android:fillColor="#e4e4e4"/>
+  <path
+      android:pathData="M312.438,574.928c7.84,-13.871 16.932,-29.288 33.181,-34.216a37.026,37.026 0,0 1,13.955 -1.441c1.482,0.128 1.112,2.412 -0.367,2.285a34.398,34.398 0,0 0,-22.272 5.892c-6.28,4.275 -11.17,10.218 -15.308,16.519 -2.535,3.861 -4.806,7.884 -7.076,11.903C313.826,577.155 311.704,576.227 312.438,574.928Z"
+      android:fillColor="#f2f2f2"/>
+  <path
+      android:pathData="M572.662,406.791L0,406.791L0,0L572.662,0Z"
+      android:fillColor="#fff"/>
+  <path
+      android:pathData="M46.648,167.076a5.58,5.58 0,0 0,0 11.16L240.038,178.236a5.58,5.58 0,0 0,0 -11.16Z"
+      android:fillColor="#d98635"/>
+  <path
+      android:pathData="M46.648,200.555a5.58,5.58 0,1 0,-0.024 11.16L240.18,211.714a5.58,5.58 0,1 0,0 -11.16Z"
+      android:fillColor="#d98635"/>
+  <path
+      android:pathData="M46.648,233.555a5.58,5.58 0,1 0,-0.024 11.16h43.556a5.58,5.58 0,1 0,0 -11.16Z"
+      android:fillColor="#d98635"/>
+  <path
+      android:pathData="M572.662,406.791L0,406.791L0,0L572.662,0ZM9.414,397.377h553.835L563.249,9.414h-553.835Z"
+      android:fillColor="#e5e5e5"/>
+  <path
+      android:pathData="M392.472,317.406a10.037,10.037 0,0 0,-0.042 20.074h95.838a10.037,10.037 0,1 0,0 -20.074Z"
+      android:fillColor="#e4e4e4"/>
+  <path
+      android:pathData="M392.472,195.859a10.037,10.037 0,0 0,-0.042 20.074h95.838a10.037,10.037 0,1 0,0 -20.074Z"
+      android:fillColor="#e4e4e4"/>
+  <path
+      android:pathData="M46.648,322.055a5.58,5.58 0,1 0,-0.024 11.16L240.18,333.214a5.58,5.58 0,1 0,0 -11.16Z"
+      android:fillColor="#e4e4e4"/>
+  <path
+      android:pathData="M46.648,355.055a5.58,5.58 0,1 0,-0.024 11.16h43.556a5.58,5.58 0,1 0,0 -11.16Z"
+      android:fillColor="#e4e4e4"/>
+  <path
+      android:pathData="M46.648,288.576a5.58,5.58 0,0 0,0 11.16L240.038,299.736a5.58,5.58 0,0 0,0 -11.16Z"
+      android:fillColor="#e4e4e4"/>
+  <path
+      android:pathData="M46.648,45.576a5.58,5.58 0,0 0,0 11.16L240.038,56.736a5.58,5.58 0,0 0,0 -11.16Z"
+      android:fillColor="#e4e4e4"/>
+  <path
+      android:pathData="M46.648,79.055a5.58,5.58 0,1 0,-0.024 11.16L240.18,90.214a5.58,5.58 0,1 0,0 -11.16Z"
+      android:fillColor="#e4e4e4"/>
+  <path
+      android:pathData="M46.648,112.055a5.58,5.58 0,1 0,-0.024 11.16h43.556a5.58,5.58 0,1 0,0 -11.16Z"
+      android:fillColor="#e4e4e4"/>
+  <path
+      android:pathData="M392.472,74.406a10.037,10.037 0,0 0,-0.042 20.074h95.838a10.037,10.037 0,1 0,0 -20.074Z"
+      android:fillColor="#e4e4e4"/>
+  <path
+      android:pathData="M412.662,84.348m-18,0a18,18 0,1 1,36 0a18,18 0,1 1,-36 0"
+      android:fillColor="#cacaca"/>
+  <path
+      android:pathData="M444.122,210.248a9.978,9.978 0,0 0,10.276 11.335l9.559,20.7 13.996,-2.741L464.045,210.499a10.032,10.032 0,0 0,-19.922 -0.251Z"
+      android:fillColor="#ffb6b6"/>
+  <path
+      android:pathData="M540.954,288.584l-39,-19.291l-27.52,-46.707l-19.236,9.305l24.964,53.957l63.064,36.405l-2.272,-33.669z"
+      android:fillColor="#3f3d56"/>
+  <path
+      android:pathData="M452.907,564.696l-12.26,-0.001l-5.833,-35.348l18.095,0.001l-0.002,35.348z"
+      android:fillColor="#ffb6b6"/>
+  <path
+      android:pathData="M456.034,576.58l-39.531,-0.001v-0.5a15.387,15.387 0,0 1,15.386 -15.386h0.001l24.144,0.001Z"
+      android:fillColor="#2f2e41"/>
+  <path
+      android:pathData="M575.282,542.21l-8.554,8.782l-29.392,-20.484l12.625,-12.962l25.321,24.664z"
+      android:fillColor="#ffb6b6"/>
+  <path
+      android:pathData="M585.977,548.262l-27.582,28.318 -0.358,-0.349a15.387,15.387 0,0 1,-0.287 -21.758l0.001,-0.001 16.846,-17.296Z"
+      android:fillColor="#2f2e41"/>
+  <path
+      android:pathData="M468.064,354.326h-8.479l0,0a236.912,236.912 0,0 0,-49.423 83.371L400.032,467.356l36.102,75.59 22.528,0.402 -25.748,-80.252 25.94,-34.4 1.308,69.152 83.492,43.763 16.71,-24.685 -67.94,-37.15 11.57,-58.82s23.279,-19.412 9.723,-49.76Z"
+      android:fillColor="#2f2e41"/>
+  <path
+      android:pathData="M579.662,277.348 L544.427,261.915 505.644,279.101s-10.982,6.247 -3.982,22.247l-42.078,52.978s0.078,0.022 52.96,16.87l47.711,-42.016a58.743,58.743 0,0 0,19.407 -51.832Z"
+      android:fillColor="#3f3d56"/>
+  <path
+      android:pathData="M573.749,236.899m-24.561,0a24.561,24.561 0,1 1,49.122 0a24.561,24.561 0,1 1,-49.122 0"
+      android:fillColor="#ffb6b6"/>
+  <path
+      android:pathData="M605.557,204.305c-5.637,-4.586 -14.379,-5.587 -20.198,-1.235 -10.149,-5.327 -23.865,-2.548 -31.14,6.309 -2.618,3.187 -4.473,7.499 -3.178,11.415 1.295,3.916 6.882,6.218 9.794,3.297l-0.052,1.203a19.495,19.495 0,0 1,-12.12 33.198q12.073,4.877 24.145,9.754c3.387,1.368 6.881,2.758 10.533,2.739s7.545,-1.793 8.962,-5.159q1.529,4.685 3.057,9.37a154.226,154.226 0,0 0,14.715 -30.337c2.437,-6.846 4.406,-14.005 4.143,-21.266C613.956,216.331 611.194,208.891 605.557,204.305Z"
+      android:fillColor="#2f2e41"/>
+  <path
+      android:pathData="M412.662,327.348m-18,0a18,18 0,1 1,36 0a18,18 0,1 1,-36 0"
+      android:fillColor="#cacaca"/>
+  <path
+      android:pathData="M468.662,205.895m-18,0a18,18 0,1 1,36 0a18,18 0,1 1,-36 0"
+      android:fillColor="#d98635"/>
+  <path
+      android:pathData="M469.122,200.248a9.978,9.978 0,0 0,10.276 11.335l9.559,20.7 13.996,-2.741L489.045,200.499a10.032,10.032 0,0 0,-19.922 -0.251Z"
+      android:fillColor="#ffb6b6"/>
+  <path
+      android:pathData="M565.954,278.584l-39,-19.291l-27.52,-46.707l-19.236,9.305l24.964,53.957l63.064,36.405l-2.272,-33.669z"
+      android:fillColor="#3f3d56"/>
+  <path
+      android:pathData="M675.658,577.348h-381a1,1 0,0 1,0 -2h381a1,1 0,0 1,0 2Z"
+      android:fillColor="#cacaca"/>
+</vector>
