+++ app/src/main/res/layout/activity_help_signalement.xml
@@ -40,13 +40,6 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/linearLayout5">
 
-        <com.google.android.material.switchmaterial.SwitchMaterial
-            android:layout_width="match_parent"
-            android:layout_height="60dp"
-            android:checked="true"
-            android:text="Tutorial"
-            android:textColor="@color/colorPrimary"/>
-
         <Button
             android:id="@+id/buttonProblemSignal"
             android:layout_width="match_parent"
