+++ app/src/main/res/layout/activity_on_boarding.xml
@@ -4,4 +4,5 @@
     android:id="@+id/pager"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".mvp.help.OnBoardingActivity"/>
+    tools:context=".mvp.help.OnBoardingActivity"
+    android:orientation="vertical"/>
