+++ app/src/main/res/xml/help.xml
@@ -7,16 +7,14 @@
         android:icon="@drawable/arrow_back_24px"
         />
 
-    <SwitchPreferenceCompat
+    <Preference
         app:icon="@drawable/ic_info_black_24dp"
         app:key="@string/setting_tutorial_key"
         app:defaultValue="@bool/setting_tutorial_default_value"
-        app:summaryOff="@string/setting_tutorial_summary_off"
-        app:summaryOn="@string/setting_tutorial_summary_on"
         app:title="@string/setting_tutorial_name" />
     <Preference
         android:title="@string/setting_help_signal_problem"
-        android:key="@string/setting_network_return_key"
+        android:key="@string/setting_help_signal_problem_key"
         android:icon="@drawable/ic_report_problem_black_24dp"
         />
 </PreferenceScreen>
